Actualizando cambios, politicas, usuarios, permisos y base de datos "condition"

+++ app/Http/Controllers/Clinic/PatientDiagnosticController.php
@@ -6,18 +6,28 @@ use App\Models\clinic\Diagnostic;
 use App\Models\clinic\Patient;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use Inertia\Inertia;
 
 
 class PatientDiagnosticController extends Controller
 {
-    
+    public function __construct(){
+        //$this->authorizeResource(Diagnostic::class); //si funciona
+    }
     public function create(int $id){
+        //$this->authorize('createDiagnostic',[request()->user(),$id]); //no funciona
+        if(request()->user()->cannot('createDiagnostic',Diagnostic::class)){ //asi porque es en $int lo arriuna
+            abort(403,'THIS ACTION IS UNAUTHORIZED. '); // es igual $this->authorize()
+        }
         return Inertia::render('Clinic/Patients/Patient_Diagnostics/CreateEditPatientDiagnostic',[
             'patient' => Patient::select(['id','name'])->get()->find($id)
         ]);
     }
     public function store(Request $request){
+        if(request()->user()->cannot('createDiagnostic',Diagnostic::class)){ //asi porque es en $int lo arriuna
+            abort(403); // es igual $this->authorize()
+        }
 
         $attributes = $request->validate([
             'patient_id' => 'required',
@@ -36,12 +46,15 @@ class PatientDiagnosticController extends Controller
     }
 
     public function edit(Diagnostic $diagnostic){
+        $this->authorize('update',$diagnostic);
         return Inertia::render('Clinic/Patients/Patient_Diagnostics/CreateEditPatientDiagnostic',[
             'patient' => Patient::select(['id','name'])->get()->find($diagnostic->patient_id),
             'patient_diagnostic' => $diagnostic
         ]);
     }
     public function update(Request $request , Diagnostic $diagnostic){
+        $this->authorize('update',$diagnostic);
+
         $attributes = $request->validate([
             'patient_id' => 'required',
             'diagnostic' => 'required|max:5000',
@@ -57,7 +70,8 @@ class PatientDiagnosticController extends Controller
 
     }
     public function destroy(Diagnostic $diagnostic){
-
+        $this->authorize('delete',$diagnostic); //no funciona ?
+        
         $diagnostic->delete();
         return back()->with([
 			'type' => 'success',
